server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["'dmytro-ustynov' <'ustynov.dmitriy@gmail.com'>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
fastapi-async-sqlalchemy = "^0.4.1"
pytube = "^15.0.0"
PyJWT = "^2.7.0"
python-decouple = "^3.8"
python-dotenv = "^1.0.0"
PyYAML = "^6.0"
pydub = "^0.25.1"
uvicorn = "^0.22.0"
bcrypt = "^4.0.1"
requests = "^2.31.0"
asyncpg = "^0.27.0"
python-multipart = "^0.0.6"
SQLAlchemy-Utils = "^0.41.1"
passlib = "^1.7.4"
aiofiles = "^23.1.0"
pydantic = {extras = ["email"], version = "^1.10.9"}
python-telegram-bot = "^20.7"
redis = "^5.0.1"
pillow = "^10.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
